move ag-service
